% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs-acceptance.R
\name{refresh_project_acceptance}
\alias{refresh_project_acceptance}
\title{Loads acceptance docs based on the yaml config file}
\usage{
refresh_project_acceptance(config = ".acceptance.yml")
}
\arguments{
\item{config}{path to the yaml configuration file}
}
\description{
Loads acceptance docs based on the yaml config file
}
