% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gs.R
\name{use_gs}
\alias{use_gs}
\title{Function that takes the Google Spreadsheet document key to save it in the working dir}
\usage{
use_gs(x, extension = "csv")
}
\arguments{
\item{x}{key of the Google Sheet document}

\item{extension}{file extension for local data, either `csv`` or `json`. defaults to csv.}
}
\description{
Function that takes the Google Spreadsheet document key to save it in the working dir
}
